azure_rag/requirements.txt
azure-identity
azure-core
azure-search-documents
azure-ai-projects
python-dotenv
requests
streamlit